requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.19
python-dotenv==1.0.0
Werkzeug==2.3.6
opencv-python==4.8.0.74
pytesseract==0.3.10
Pillow==10.0.0
numpy==1.24.3
ultralytics==8.0.196
easyocr==1.7.1
torch==2.0.1
torchvision==0.15.2
